requirements.txt
humanize
requests
six
rsa
bs4
schematics
humanfriendly
gtts
googletrans
wikiapi
null
wikipedia
python-dotenv
pyngrok
thrift